refactor: cleanup mess in the model directory
