+++ backend/app/src/main/java/backend/model/pair/PairResponse.java
@@ -1,4 +1,4 @@
-package backend.model;
+package backend.model.pair;
 
 import org.javatuples.Pair;
